async_api/etl/requirements.txt
backoff==2.2.1
elasticsearch==7.17
flake8==6.0.0
loguru==0.6.0
psycopg2-binary==2.9.5
psycopg2==2.9.5
pydantic==1.10.5
python-dotenv==1.0.0
